Make common public methods for network utils
